refactor: Rename `lib/plugins/aws/lib/monitorStack.js`
