+++ lib/plugins/aws/deploy/index.js
@@ -5,7 +5,7 @@ const writeServiceOutputs = require('../../../cli/write-service-outputs');
 const extendedValidate = require('./lib/extendedValidate');
 const setBucketName = require('../lib/setBucketName');
 const checkForChanges = require('./lib/checkForChanges');
-const monitorStack = require('../lib/monitorStack');
+const monitorStack = require('../lib/monitor-stack');
 const checkIfBucketExists = require('../lib/check-if-bucket-exists');
 const getCreateChangeSetParams = require('../lib/get-create-change-set-params');
 const getExecuteChangeSetParams = require('../lib/get-execute-change-set-params');
