+++ lib/plugins/aws/lib/monitor-stack.js
