+++ lib/plugins/aws/remove/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const validate = require('../lib/validate');
-const monitorStack = require('../lib/monitorStack');
+const monitorStack = require('../lib/monitor-stack');
 const checkIfBucketExists = require('../lib/check-if-bucket-exists');
 const emptyS3Bucket = require('./lib/bucket');
 const removeStack = require('./lib/stack');
