+++ lib/plugins/aws/rollback.js
@@ -2,7 +2,7 @@
 const validate = require('./lib/validate');
 const setBucketName = require('./lib/setBucketName');
 const updateStack = require('./lib/updateStack');
-const monitorStack = require('./lib/monitorStack');
+const monitorStack = require('./lib/monitor-stack');
 const waitForChangeSetCreation = require('./lib/wait-for-change-set-creation');
 const getCreateChangeSetParams = require('./lib/get-create-change-set-params');
 const getExecuteChangeSetParams = require('./lib/get-execute-change-set-params');
